backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
google-cloud-vision==3.4.4
openai==1.3.0
transformers==4.35.0
torch==2.1.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23 